Add pathPrefixTree.
