+++ ch/ch_test.go
@@ -32,16 +32,13 @@ import (
 	"os"
 	"os/user"
 	"path/filepath"
-	"sort"
 	"strconv"
-	"strings"
 	"syscall"
 	"testing"
 	"time"
 
 	"github.com/inconshreveable/log15"
 	. "github.com/smartystreets/goconvey/convey"
-	suffix "github.com/spacewander/go-suffix-tree"
 )
 
 const longBasename = "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA" +
@@ -533,350 +530,4 @@ func getFilePermissions(t *testing.T, path string) string {
 	So(err, ShouldBeNil)
 
 	return info.Mode().Perm().String()
-}
-
-var prefixBenchmarkPaths []string
-
-const prefixBenchmarkNumPaths = 1000000
-
-var prefixBenchmarkPrefixes []string
-
-const prefixBenchmarkNumPrefixes = 1000
-
-func prefixBenchmarkSetup() {
-	if prefixBenchmarkPaths == nil {
-		makePrefixBenchmarkPaths()
-	}
-
-	if prefixBenchmarkPrefixes == nil {
-		makePrefixBenchmarkPrefixes()
-	}
-}
-
-func makePrefixBenchmarkPaths() {
-	prefixBenchmarkPaths = make([]string, 0, prefixBenchmarkNumPaths)
-
-	makePrefixPath("/", 10, 7, func(path string) {
-		if len(prefixBenchmarkPaths) != prefixBenchmarkNumPaths {
-			prefixBenchmarkPaths = append(prefixBenchmarkPaths, filepath.Join(path, "file.txt"))
-		}
-	})
-
-}
-
-func makePrefixPath(path string, max, remaining int, cb func(string)) {
-	if remaining == 0 {
-		cb(path)
-
-		return
-	}
-
-	for i := 0; i < max; i++ {
-		iDir := strconv.Itoa(i)
-
-		makePrefixPath(filepath.Join(path, iDir), max, remaining-1, cb)
-	}
-}
-
-func makePrefixBenchmarkPrefixes() {
-	prefixBenchmarkPrefixes = make([]string, 0, prefixBenchmarkNumPrefixes)
-
-	cb := func(path string) {
-		prefixBenchmarkPrefixes = append(prefixBenchmarkPrefixes, path)
-	}
-
-	makePrefixPath("/", 5, 8, cb)
-	makePrefixPath("/", 5, 6, cb)
-	makePrefixPath("/", 5, 5, cb)
-	makePrefixPath("/", 5, 4, cb)
-	makePrefixPath("/", 5, 3, cb)
-
-}
-
-func loopPrefixes() {
-	for _, path := range prefixBenchmarkPaths {
-		for _, prefix := range prefixBenchmarkPrefixes {
-			if strings.HasPrefix(path, prefix) {
-				break
-			}
-		}
-	}
-}
-
-func BenchmarkPrefixMap(b *testing.B) {
-	prefixBenchmarkSetup()
-
-	prefixMap := make(map[string]bool, prefixBenchmarkNumPrefixes)
-
-	for _, prefix := range prefixBenchmarkPrefixes {
-		prefixMap[prefix] = true
-	}
-
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		for _, path := range prefixBenchmarkPaths {
-			checkPathParents(prefixMap, path)
-		}
-	}
-}
-
-func checkPathParents(prefixMap map[string]bool, path string) bool {
-	dir := path
-
-	for {
-		if prefixMap[dir] {
-			return true
-		}
-
-		dir = filepath.Dir(dir)
-		if dir == "." || dir == "/" {
-			return false
-		}
-	}
-}
-
-func BenchmarkPrefixTree(b *testing.B) {
-	prefixBenchmarkSetup()
-
-	tree := splitPrefixesToTree()
-
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		checkPathsWithTree(tree)
-	}
-}
-
-type prefixTree struct {
-	children map[string]*prefixTree
-	leaf     string
-}
-
-func newPrefixTree() *prefixTree {
-	return &prefixTree{
-		children: make(map[string]*prefixTree),
-	}
-}
-
-func (p *prefixTree) add(directory string) {
-	for dir, rest := splitPath(directory[1:]); dir != ""; dir, rest = splitPath(rest) {
-		p = p.child(dir)
-	}
-
-	p.leaf = directory
-}
-
-func splitPath(path string) (string, string) {
-	pos := strings.IndexByte(path, filepath.Separator)
-
-	if pos == -1 {
-		return path, ""
-	}
-
-	return path[:pos], path[pos+1:]
-}
-
-func (p *prefixTree) child(directory string) *prefixTree {
-	tree, exists := p.children[directory]
-	if !exists {
-		tree = newPrefixTree()
-		p.children[directory] = tree
-	}
-
-	return tree
-}
-
-func (p *prefixTree) getPrefixForPath(path string) (string, bool) {
-	for dir, rest := splitPath(path[1:]); dir != ""; dir, rest = splitPath(rest) {
-		tree, found := p.children[dir]
-		if !found {
-			if p.leaf != "" {
-				return p.leaf, true
-			}
-
-			break
-		}
-
-		p = tree
-	}
-
-	return "", false
-}
-
-func splitPrefixesToTree() *prefixTree {
-	tree := newPrefixTree()
-
-	for _, prefix := range prefixBenchmarkPrefixes {
-		tree.add(prefix)
-	}
-
-	return tree
-}
-
-func checkPathsWithTree(tree *prefixTree) {
-	for _, path := range prefixBenchmarkPaths {
-		tree.getPrefixForPath(path)
-	}
-}
-
-func BenchmarkPrefixSuffixTree(b *testing.B) {
-	prefixBenchmarkSetup()
-
-	tree := generateSuffixTree()
-
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		checkPathsWithSuffixTree(tree)
-	}
-}
-
-func generateSuffixTree() *suffix.Tree {
-	tree := suffix.NewTree()
-
-	for _, prefix := range prefixBenchmarkPrefixes {
-		p := prefix
-		tree.Insert([]byte(prefix), &p)
-	}
-
-	return tree
-}
-
-func checkPathsWithSuffixTree(tree *suffix.Tree) {
-	for _, path := range prefixBenchmarkPaths {
-		tree.LongestSuffix([]byte(path))
-	}
-}
-
-func BenchmarkPrefixDepth(b *testing.B) {
-	prefixBenchmarkSetup()
-
-	d := generateDepth()
-
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		for _, path := range prefixBenchmarkPaths {
-			getDepthPrefix(d, path)
-		}
-	}
-}
-
-type Depth struct {
-	prefixes map[string]struct{}
-	depths   []int
-}
-
-func generateDepth() *Depth {
-	d := &Depth{
-		prefixes: make(map[string]struct{}),
-	}
-
-	counts := make(map[int]struct{})
-
-	for _, path := range prefixBenchmarkPrefixes {
-		d.prefixes[path] = struct{}{}
-		counts[strings.Count(path, "/")] = struct{}{}
-	}
-
-	for count := range counts {
-		d.depths = append(d.depths, count)
-	}
-
-	sort.Sort(sort.Reverse(sort.IntSlice(d.depths)))
-
-	return d
-}
-
-func getNthSep(path string, sep byte, n int) string {
-	lastPos := 0
-	for i := 0; i < n; i++ {
-		pos := strings.IndexByte(path[lastPos+1:], sep)
-
-		if pos == -1 {
-			return ""
-		}
-
-		lastPos += pos + 1
-	}
-
-	return path[:lastPos]
-}
-
-func getDepthPrefix(d *Depth, path string) string {
-	for _, count := range d.depths {
-		cutPath := getNthSep(path, filepath.Separator, count)
-
-		_, ok := d.prefixes[cutPath]
-		if ok {
-			return cutPath
-		}
-	}
-
-	return ""
-}
-
-// go test ./ch -run TestPrefixMatchers
-// .
-func TestPrefixMatchers(t *testing.T) {
-	prefixBenchmarkSetup()
-
-	Convey("Map works", t, func() {
-		prefixMap := make(map[string]bool, prefixBenchmarkNumPrefixes)
-
-		for _, prefix := range prefixBenchmarkPrefixes {
-			prefixMap[prefix] = true
-		}
-
-		found := checkPathParents(prefixMap, "/9/file.txt")
-		So(found, ShouldBeFalse)
-
-		found = checkPathParents(prefixMap, "/0/file.txt")
-		So(found, ShouldBeFalse)
-
-		found = checkPathParents(prefixMap, filepath.Join(prefixBenchmarkPrefixes[0], "file.txt"))
-		So(found, ShouldBeTrue)
-	})
-
-	Convey("Depth works", t, func() {
-		d := generateDepth()
-
-		found := getDepthPrefix(d, "/9/file.txt")
-		So(found, ShouldBeBlank)
-
-		found = getDepthPrefix(d, "/0/file.txt")
-		So(found, ShouldBeBlank)
-
-		found = getDepthPrefix(d, filepath.Join(prefixBenchmarkPrefixes[0], "file.txt"))
-		So(found, ShouldEqual, prefixBenchmarkPrefixes[0])
-	})
-
-	Convey("Tree works", t, func() {
-		tree := splitPrefixesToTree()
-
-		_, found := tree.getPrefixForPath("/9/file.txt")
-		So(found, ShouldBeFalse)
-
-		_, found = tree.getPrefixForPath("/0/file.txt")
-		So(found, ShouldBeFalse)
-
-		prefix, found := tree.getPrefixForPath(filepath.Join(prefixBenchmarkPrefixes[0], "file.txt"))
-		So(found, ShouldBeTrue)
-		So(prefix, ShouldEqual, prefixBenchmarkPrefixes[0])
-	})
-
-	// Convey("Suffix tree works", t, func() {
-	// 	tree := generateSuffixTree()
-
-	// 	_, _, found := tree.LongestSuffix([]byte("/9/file.txt"))
-	// 	So(found, ShouldBeFalse)
-
-	// 	_, _, found = tree.LongestSuffix([]byte("/0/file.txt"))
-	// 	So(found, ShouldBeFalse)
-
-	// 	k, _, found := tree.LongestSuffix([]byte(filepath.Join(prefixBenchmarkPrefixes[0], "file.txt")))
-	// 	So(found, ShouldBeTrue)
-	// 	So(k, ShouldResemble, []byte(prefixBenchmarkPrefixes[0]))
-	// })
 }
